Add About section for mobile devices. Move fetch data from About.jsx to App.jsx
