+++ styles/main.css
@@ -173,6 +173,20 @@ p {
   display: none;
 }
 
+.bio-mobile {
+  margin: 3rem 0;
+}
+.bio-mobile h2 {
+  position: relative;
+}
+.bio-mobile h2::before {
+  content: url("/assets/icons/rocket.svg");
+  margin-right: 0.525rem;
+}
+.bio-mobile p {
+  margin: 0.5rem 0;
+}
+
 @media screen and (min-width: 640px) {
   .about .about-icons li a span {
     display: block;
@@ -231,6 +245,9 @@ p {
   .about .about-social .about-icons .list-item span {
     margin-left: 0.75rem;
     font-weight: 500;
+  }
+  .bio-mobile {
+    display: none;
   }
 }
 .btn-cta {
